get target account id from aws_caller_identity
This is a breaking change which removes the `csw_target_account_id`
variable and instead gets the target account id from the
`aws_caller_identity` data source [1].  The advantage is that it's
less configuration for the user to pass in.

I considered making this a graceful change which simply adds a default
value to the csw_target_account_id so that existing clients continue
to work but I think it's probably easiest to just make the change and
update consumers.  However, if people are depending on `master` then
their terraform builds may suddenly break.

[1] https://www.terraform.io/docs/providers/aws/d/caller_identity.html